migrated model related functions to experiment.rb [fixes: SCI-31]
